fix: salario from double to decimal

+++ Models/Empleado.cs
@@ -11,9 +11,9 @@ public class Empleado : Persona
     public Guid Id = Guid.NewGuid();
     public string NumeroDeIdentificacion { get; set; }
     public string Posicion { get; set; }
-    public double Salario { get; set; }
+    public decimal Salario { get; set; }
 
-    public Empleado(Guid id, string nombre, string apellidos, string numeroDeIdentificacion, int edad, string posicion, double salario) : base ( nombre , apellidos, edad)
+    public Empleado(Guid id, string nombre, string apellidos, string numeroDeIdentificacion, int edad, string posicion, decimal salario) : base ( nombre , apellidos, edad)
     {
         Id = id;
         Nombre = nombre;
@@ -24,9 +24,9 @@ public class Empleado : Persona
         Salario = salario;
     }
 
-    private double CalcularBonificacion(double Salario)
+    private decimal CalcularBonificacion(decimal Salario)
     {
-        double bonificacion = Salario +(Salario / 100* 10);
+        decimal bonificacion = Salario +(Salario / 100* 10);
         
         return bonificacion;
     }
